Fix commented Contents for precise value.
ONE-DCO-1.0-Signed-off-by: KiDeuk Bang <rrstrous@nate.com>

+++ compiler/luci-interpreter/src/kernels/TestUtils.h
@@ -96,8 +96,8 @@ template <typename T> std::pair<float, int32_t> quantizationParams(float f_min, 
   float scale = 0;
   const T qmin = std::numeric_limits<T>::min();
   const T qmax = std::numeric_limits<T>::max();
-  const float qmin_double = qmin;
-  const float qmax_double = qmax;
+  const double qmin_double = qmin;
+  const double qmax_double = qmax;
   // 0 should always be a representable value. Let's assume that the initial
   // min,max range contains 0.
   assert(f_max >= 0);
@@ -123,16 +123,16 @@ template <typename T> std::pair<float, int32_t> quantizationParams(float f_min, 
   // The arithmetic error on the zero point computed from either pair
   // will be roughly machine_epsilon * (sum of absolute values of terms)
   // so we want to use the variant that adds the smaller terms.
-  const float zero_point_from_min = qmin_double - f_min / scale;
-  const float zero_point_from_max = qmax_double - f_max / scale;
+  const double zero_point_from_min = qmin_double - f_min / scale;
+  const double zero_point_from_max = qmax_double - f_max / scale;
 
-  const float zero_point_from_min_error = std::abs(qmin_double) + std::abs(f_min / scale);
+  const double zero_point_from_min_error = std::abs(qmin_double) + std::abs(f_min / scale);
 
-  const float zero_point_from_max_error = std::abs(qmax_double) + std::abs(f_max / scale);
+  const double zero_point_from_max_error = std::abs(qmax_double) + std::abs(f_max / scale);
 
-  const float zero_point_double = zero_point_from_min_error < zero_point_from_max_error
-                                      ? zero_point_from_min
-                                      : zero_point_from_max;
+  const double zero_point_double = zero_point_from_min_error < zero_point_from_max_error
+                                       ? zero_point_from_min
+                                       : zero_point_from_max;
 
   // Now we need to nudge the zero point to be an integer
   // (our zero points are integer, and this is motivated by the requirement
@@ -160,7 +160,7 @@ template <typename T> std::pair<float, int32_t> quantizationParams(float f_min, 
   assert(qmin <= nudged_zero_point);
   zero_point = nudged_zero_point;
   // finally, return the values
-  return {scale, zero_point};
+  return {static_cast<float>(scale), zero_point};
 }
 
 } // namespace testing
